SERVER-8203 Fix Windows compile warning
warning C4146: unary minus applied to unsigned type, result still unsigned

+++ src/mongo/db/repl/bgsync.cpp
@@ -431,7 +431,7 @@ namespace replset {
         // and queued for application already
         BSONObj op = _buffer.blockingPop();
         bufferCountGauge.increment(-1);
-        bufferSizeGauge.increment(-getSize(op));
+        bufferSizeGauge.increment(-static_cast<int64_t>(getSize(op)));
     }
 
     bool BackgroundSync::isStale(OplogReader& r, BSONObj& remoteOldestOp) {
